clarify responsibility
I've removed confusing language about PEOs versus CI employees, as it is not necessary for the CoCc to understand the distinction or to know which Core Team members are under each type of employment. I've clarified that the responsibility is for the CoCc to notify the Core Team liaison in any case of doubt as to whether an incident involves Core Team member(s) and harassment.